+++ CHANGELOG.md
@@ -2,4 +2,4 @@
 
 ## 0.1.0.0
 
-* Build recursive directory trees and avoid symlink cycles.
+* Build recursive directory trees and try to avoid symlink cycles. Includes functions to walk a directory tree, filter it, and display it.